Fix slide_rois and preview_from_queries
Fixed a bug in slide_rois that made the patches not correctly aligned with the ancestors and in fact overwriting each other in the patches.csv file.
The same bug was present in slide_rois_no_image.

I had a shift with the preview_from_queries function that made the resulting patches not match what was visible on the thumbnail. The problem seems to be solved with this version.
Also wouldn't the idea be to do startx = numpy.clip(x, 0, thumb_size.x - 1) instead of startx = min(x, thumb_size.x - 1) ?

And in util.types I replaced the N with * which allow the images to not have same width and height :
isinstance(numpy.ones(3, 2),  NDArray[Shape["N, N"], numpy.uint8]) return False
but
isinstance(numpy.ones(3, 2),  NDArray[Shape["*, *"], numpy.uint8]) return True