Updated preview_from_queries
I updated the .github/workflows/tests_run.yml with the one from the master branch.

I added a check in the visu function to not make the program crash if the function is used with an empty patch list.
I also replaced the int with round in the function which seems to give slihtly better results when used with thickness=1.